Get last plot projection

'tissot_map()' will add the [world] coastline to any map.

tissot_map(..., add = TRUE)

tissot_abline(lambda, phi = NULL, ..., proj.in = NULL)

tissot_get_proj()

Arguments

...
graphical parameters for [lines()] if 'add = TRUE', or for [plot()] if 'add = FALSE'

add
logical, default 'TRUE' add to existing plot or create new

lambda
longitude at which to draw a vertical line

phi
latitude at which to draw a horizontal line

proj.in
projection for expert use

Value

'tissot_map()' returns the internal world map data (projected if one is current) as a matrix

'tissot_abline()' called for its side effect of drawing on the plot

'tissot_get_proj()' returns the value of the current projection, or NULL

Details

'tissot_get_proj()' When the indicatrix is plotted it registers its projection. This string can be obtained with this getter function.

'tissot_abline()' will draw a vertical and horizontal line at a give longitude latitude (where they intersect is the actual lon,lat location)